feat:加入初始化钩子
